+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Fixed
 - Fixed issue of autostarted agents not being restarted on environment setting change (#2049)
 - Log primary for agent correctly in the database when pausing/unpausing agents (#2079)
+- Cancel scheduled deploy operations of an agent when that agent is paused (#2077)
+- Fix agent-names config type (#2071)
 - Ensure the internal agent is always present in the autostart_agent_map of auto-started agents (#2101)
 
 ## Added